+++ example/country.dart
@@ -9,7 +9,9 @@ class Country {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is Country && runtimeType == other.runtimeType && name == other.name;
+      other is Country &&
+          runtimeType == other.runtimeType &&
+          name == other.name;
 
   @override
   int get hashCode => name.hashCode;