+++ example/student.dart
@@ -2,18 +2,12 @@ class Student {
   final int id;
   final String name;
 
-  Student({
-    required this.id,
-    required this.name,
-  });
+  Student({required this.id, required this.name});
 }
 
 class Hello {
   final int id;
   final String name;
 
-  Hello({
-    required this.id,
-    required this.name,
-  });
+  Hello({required this.id, required this.name});
 }