+++ lib/riverpod_repo.dart
@@ -3,7 +3,5 @@ import 'package:riverpod_repo/src/riverpod_repo_generator.dart';
 import 'package:source_gen/source_gen.dart';
 
 /// Generates the Riverpod providers based on the repository.
-Builder generateRiverpodRepo(BuilderOptions options) => LibraryBuilder(
-      RiverPodRepoGenerator(),
-      generatedExtension: '.repo.dart',
-    );
+Builder generateRiverpodRepo(BuilderOptions options) =>
+    LibraryBuilder(RiverPodRepoGenerator(), generatedExtension: '.repo.g.dart');